+++ MainProject/SpringBootTurialVip/src/main/java/com/example/SpringBootTurialVip/shoprepository/ProductRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 	List<Product> findByIsActiveTrue();
 
@@ -27,4 +27,6 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
 			Pageable pageable);
 
 	List<Product> findByTitle(String title);
+
+
 }
